==> models/resized_data_case/3. KNN Naive Models.ipynb <==
# Note

* Use the extracted features from pre-trained Inception-V3 model
* Fit KNN model directly on the whole training set
* Find first 20 neighbors for all the test images

import warnings
warnings.simplefilter('ignore')

import numpy as np
import pandas as pd
from scipy.stats import mode
from sklearn.neighbors import NearestNeighbors

%matplotlib inline

# Load Features and Labels

train_feature = np.load('./data/knn/train_feature.npy')
test_feature = np.load('./data/knn/test_feature.npy')

train_df = pd.read_csv('./data/knn/train.csv')
test_df = pd.read_csv('./data/knn/test.csv')

print('Train:\t', train_feature.shape, train_df.shape)
print('Test:\t', test_feature.shape, test_df.shape)

train_df.head()

# 0. Random Guess

# Helper function
def accuracy(true_label, prediction, top=1):
    """ function to calculate the prediction accuracy """
    prediction = prediction[:, :top]
    count = 0
    for i in range(len(true_label)):
        if true_label[i] in prediction[i]:
            count += 1
            
    return count / len(true_label)

### Random guessing

seed = 42
random_guess = np.random.choice(train_df['landmark_id'].values, 
                                size=(len(test_df), 20), replace=True)

print('Top  1 accuracy:\t', accuracy(test_df['landmark_id'].values, random_guess, top=1))
print('Top  5 accuracy:\t', accuracy(test_df['landmark_id'].values, random_guess, top=5))
print('Top 10 accuracy:\t', accuracy(test_df['landmark_id'].values, random_guess, top=10))
print('Top 20 accuracy:\t', accuracy(test_df['landmark_id'].values, random_guess, top=20))

# 1. Implement KNN Model (unnormalized features)

# Implement KNN model
knn = NearestNeighbors(n_neighbors=20, algorithm='auto', leaf_size=30, 
                       metric='minkowski', p=2, n_jobs=-1)
knn.fit(train_feature)

# Search the first 20 neighbors
distance, neighbor_index = knn.kneighbors(test_feature, return_distance=True)

# Save the results
np.save('./result/knn_naive_distance.npy', distance)
np.save('./result/knn_naive_neighbor_index.npy', neighbor_index)

## Search Neighbors

knn_distance = np.load('./result/knn_naive_distance.npy')
knn_neighbor = np.load('./result/knn_naive_neighbor_index.npy')

# Get the first 20 neighbors
predictions = []
for neighbors in knn_neighbor:
    predictions.append(train_df.loc[neighbors]['landmark_id'].values)

predictions = np.array(predictions)

## Calculate Accuracy

### Mode accuracy

prediction_mode = mode(predictions, axis=1)
print('Mode accuracy:\t', accuracy(test_df['landmark_id'].values, prediction_mode[0], top=1))

### Top 1 accuracy

print('Top  1 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=1))
print('Top  5 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=5))
print('Top 10 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=10))
print('Top 20 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=20))

# 2. Implement KNN Model (normalized features)

# Normalize features
train_norm = np.linalg.norm(train_feature, axis=1, keepdims=True)
test_norm = np.linalg.norm(test_feature, axis=1, keepdims=True)
train_feature = train_feature / train_norm
test_feature = test_feature / test_norm

# Implement KNN model
knn = NearestNeighbors(n_neighbors=20, algorithm='auto', leaf_size=30, 
                       metric='minkowski', p=2, n_jobs=-1)
knn.fit(train_feature)

# Search the first 20 neighbors
distance, neighbor_index = knn.kneighbors(test_feature, return_distance=True)

# Save the results
np.save('./result/knn_naive_distance_normalized.npy', distance)
np.save('./result/knn_naive_neighbor_index_normalized.npy', neighbor_index)

## Search Neighbors

knn_distance = np.load('./result/knn_naive_distance_normalized.npy')
knn_neighbor = np.load('./result/knn_naive_neighbor_index_normalized.npy')

# Get the first 20 neighbors
predictions = []
for neighbors in knn_neighbor:
    predictions.append(train_df.loc[neighbors]['landmark_id'].values)

predictions = np.array(predictions)

## Calculate Accuracy

### Mode accuracy

prediction_mode = mode(predictions, axis=1)
print('Mode accuracy:\t', accuracy(test_df['landmark_id'].values, prediction_mode[0], top=1))

### Top k accuracy

print('Top  1 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=1))
print('Top  5 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=5))
print('Top 10 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=10))
print('Top 20 accuracy:\t', accuracy(test_df['landmark_id'].values, predictions, top=20))
